Create a separate USB module in the capsules.
