+++ boards/imix/src/imix_components/usb.rs
@@ -23,9 +23,9 @@ pub struct UsbComponent {
     board_kernel: &'static kernel::Kernel,
 }
 
-type UsbDevice = capsules::usb_user::UsbSyscallDriver<
+type UsbDevice = capsules::usb::usb_user::UsbSyscallDriver<
     'static,
-    capsules::usbc_client::Client<'static, sam4l::usbc::Usbc<'static>>,
+    capsules::usb::usbc_client::Client<'static, sam4l::usbc::Usbc<'static>>,
 >;
 
 impl UsbComponent {
@@ -45,18 +45,18 @@ impl Component for UsbComponent {
 
         // Configure the USB controller
         let usb_client = static_init!(
-            capsules::usbc_client::Client<'static, sam4l::usbc::Usbc<'static>>,
-            capsules::usbc_client::Client::new(&sam4l::usbc::USBC)
+            capsules::usb::usbc_client::Client<'static, sam4l::usbc::Usbc<'static>>,
+            capsules::usb::usbc_client::Client::new(&sam4l::usbc::USBC)
         );
         sam4l::usbc::USBC.set_client(usb_client);
 
         // Configure the USB userspace driver
         let usb_driver = static_init!(
-            capsules::usb_user::UsbSyscallDriver<
+            capsules::usb::usb_user::UsbSyscallDriver<
                 'static,
-                capsules::usbc_client::Client<'static, sam4l::usbc::Usbc<'static>>,
+                capsules::usb::usbc_client::Client<'static, sam4l::usbc::Usbc<'static>>,
             >,
-            capsules::usb_user::UsbSyscallDriver::new(
+            capsules::usb::usb_user::UsbSyscallDriver::new(
                 usb_client,
                 self.board_kernel.create_grant(&grant_cap)
             )
