+++ capsules/src/lib.rs
@@ -54,9 +54,6 @@ pub mod temperature;
 pub mod tmp006;
 pub mod tsl2561;
 pub mod usb;
-pub mod usb_user;
-pub mod usbc_client;
-pub mod usbc_client_ctrl;
 pub mod virtual_alarm;
 pub mod virtual_flash;
 pub mod virtual_i2c;
