+++ capsules/src/usb/descriptors.rs
