+++ capsules/src/usb/mod.rs
@@ -1,0 +1,4 @@
+pub mod descriptors;
+pub mod usb_user;
+pub mod usbc_client;
+pub mod usbc_client_ctrl;
